fix: 双方向の制御文字を落とす／id フィールド名も引用する／公開の可否はルールが読む文書から
Codex の「正規表現の範囲が不正」は誤り — u007f-u009f は逆順ではなく、
モジュールは読み込まれる（tsx で quoted() を直接呼び、spec 43 本が通る）。
ただし指した行には本物の穴があった: u200b-u200f は bidi の「マーク」しか
覆わず、u202a-u202e と u2066-u2069 —— 後続を並べ替えるもの —— が抜けていた。
引用値が、それが含んでいないテキストとして描画されうる。範囲を広げ、
制御文字が潰れることを spec で固定した（これはこのモジュールが parse できる
ことの証拠でもある）。

Codex P1「id フィールド名が引用されていない」: missingIdField が返すのも
宣言の言葉で、そのまま agent に届く。引用した。

Codex P2「公開の可否を authorization の文書から導いていない」: publish は
config/public を apps/{aid} より先に書くので、途中で止まった run は新しい
意図と生きている認可を別々の文書に残す。ルールが読むのは apps/{aid}.public
なので、読めるならそちらを見る。読めない読者（コレクション単位のロール）
のために config/public、最後に slug の予約。

spec が 600 行（コメント除く）を超えたので読みと書きに割り、fake は
test/support/participateHarness.ts に出した。vi.mock はファイル単位で
hoist されるので共有できるのは「mock を作るもの」と可変の bag だけで、
各 spec は vi.hoisted で作って渡す。
