+++ test/server/backends/sharedAppParticipate.spec.ts
@@ -1,339 +1,61 @@
 // @vitest-environment node
 //
-// TAKING PART in an app somebody else published.
+// READING somebody else's app: what it is, what you may do in it, and what its records say.
 //
-// The property under test is not the one `sharedAppPreviewIntent.spec.ts` pins. There the author is
-// the app's OWNER and the host must not be looser than production; here every read and write goes
-// out as the signed-in reader, so the deployed rules answer for exactly the right person and the
-// host's judgement buys a NAMED refusal rather than a permission.
+// TAKING PART in an app somebody else published. The property under test is not the one
+// `sharedAppPreviewIntent.spec.ts` pins: there the author is the app's OWNER and the host must not
+// be looser than production, while here every read and write goes out as the signed-in reader, so
+// the deployed rules answer for exactly the right person and the host's judgement buys a NAMED
+// refusal rather than a permission.
 //
-// So what these tests hold is the other half:
+// What this half holds: the declaration is read back off FIRESTORE with no repository involved; a
+// read that BROKE is never reported as a boundary the rules drew; and every string the app's author
+// wrote arrives quoted, because it is data and the report around it is not.
 //
-//   - the declaration is read back off FIRESTORE, with no repository involved at all — an app this
-//     machine has never held is fully usable from its published documents
-//   - the pairs the rules read with `getAfter()` go out in one batch, spelled as mulmoserver spells
-//     them (the mail id especially: `{cid}_{itemId}_{template}` is rebuilt by the rules, and a
-//     divergence queues a document nothing can send)
-//   - a submission is reported as a record written and never as a place held (principle 3 — the
-//     rules cannot count rows, so capacity is derived from order and this tool must not claim one)
+// The fake Firestore and the app published into it are `test/support/participateHarness.ts`, shared
+// with the sibling file: `vi.mock` is hoisted per FILE, so the mock itself cannot be — what travels
+// is what it is made of, and the mutable bag it answers from.
 import { describe, it, expect, beforeAll, beforeEach, vi } from "vitest";
-import { appViewTierPath, viewConfigDocId } from "@receptron/sharedapp";
-import { setFirestoreAccessor, setSharedCollectionsSupport, type FirestoreDoc, type FirestoreDocs } from "@mulmoclaude/core/collection/server";
-import { useSharedApp } from "../../../server/infra/use-shared-app-tool.js";
+import { viewConfigDocId } from "@receptron/sharedapp";
+import { setFirestoreAccessor, setSharedCollectionsSupport } from "@mulmoclaude/core/collection/server";
 import { chmodSync, writeFileSync } from "node:fs";
 import path from "node:path";
+import { useSharedApp } from "../../../server/infra/use-shared-app-tool.js";
+import { AID, bookingsPath, DEFAULT_ROWS, freshBag, ME, publishApp, slotsPath, type Bag } from "../../support/participateHarness.js";
 import { makeTempDir } from "../../support/tempDir";
 
-const AID = "app-sakura";
-/** The tool's own default, restated so the assertion below says what it is checking. */
-const DEFAULT_ROWS = 50;
-const ME = { uid: "uid-me", email: "me@example.com" };
+// CREATED WITH `vi.hoisted` because the mock factory below is hoisted above the imports: a plain
+// `const` here would still be in its temporal dead zone when the factory first runs.
+const bag = vi.hoisted(
+  () =>
+    ({
+      batched: [],
+      batchFails: false,
+      batchRefusals: 0,
+      batchBreaks: 0,
+      capped: [],
+      breakQuery: new Set<string>(),
+      denyQuery: new Set<string>(),
+      queryable: new Map(),
+    }) as unknown as Bag,
+);
 
-const batched: string[] = [];
-let batchFails = false;
-/** How many commits refuse before the rest succeed — the rules saying no to one tier's attempt. */
-let batchRefusals = 0;
-/** How many commits FAIL without a permission code — a blip, which must not be read as a refusal. */
-let batchBreaks = 0;
-
-vi.mock("firebase/firestore", () => ({
-  collection: (_db: unknown, collectionPath: string) => ({ collectionPath }),
-  serverTimestamp: () => ({ __serverTimestamp: true }),
-  doc: (parent: { collectionPath: string }, docId: string) => ({ path: `${parent.collectionPath}/${docId}` }),
-  // The own-row query, recorded as a description rather than run: what matters is that the field
-  // and the value came from the DECLARATION and the session, not from anything a caller passed.
-  // Constraints arrive in whatever combination the caller built: a whole-collection read is a cap
-  // alone, an own-row read is a `where` and a cap. Sorted out here rather than by position, so the
-  // fake cannot quietly accept a query that forgot one.
-  query: (source: { collectionPath: string }, ...constraints: ({ field: string; value: unknown } | { rows: number })[]) => {
-    const clause = constraints.find((entry): entry is { field: string; value: unknown } => "field" in entry);
-    const cap = constraints.find((entry): entry is { rows: number } => "rows" in entry);
-    if (cap === undefined) throw new Error("a query was built with no row cap");
-    capped.push(cap.rows);
-    return { ...source, clause, cap };
-  },
-  // The real `where` takes a FieldPath here, whose `segments` are the literal key. The fake reads
-  // it the same way, so a host that went back to passing a bare string would query a nested path
-  // and this file would notice.
-  FieldPath: class {
-    segments: string[];
-    constructor(...segments: string[]) {
-      this.segments = segments;
-    }
-  },
-  where: (field: { segments: string[] }, _op: string, value: unknown) => ({ field: field.segments.join("."), value }),
-  limit: (rows: number) => ({ rows }),
-  getDocs: (asked: { collectionPath: string; clause?: { field: string; value: unknown }; cap: { rows: number } }) => {
-    // Refused only WITHOUT a where clause — which is the shape the rules actually take: listing a
-    // collection people submit to is denied, while a query narrowed to the reader's own rows is
-    // the one a participant's page issues and is allowed.
-    if (breakQuery.has(asked.collectionPath)) return Promise.reject(new Error("unavailable"));
-    if (denyQuery.has(asked.collectionPath) && asked.clause === undefined) {
-      return Promise.reject(Object.assign(new Error("Missing or insufficient permissions."), { code: "permission-denied" }));
-    }
-    const rows = queryable.get(asked.collectionPath) ?? [];
-    const clause = asked.clause;
-    return Promise.resolve({
-      docs: rows
-        .filter((row) => clause === undefined || row.data[clause.field] === clause.value)
-        // The fake honours the cap, so a host that stopped passing one would hand back more rows
-        // than it asked for and the assertion would notice.
-        .slice(0, asked.cap.rows)
-        .map((row) => ({ id: row.id, data: () => row.data })),
-    });
-  },
-  // The transaction the mirrored create goes through. It is not a batch with a nicer name: the
-  // `get` inside it is what the commit is re-run against, which is what makes claiming a slot
-  // atomic. The fake answers it from the same store, so a host that stopped reading would claim a
-  // slot somebody already holds and the test would notice.
-  runTransaction: async (_db: unknown, body: (tx: unknown) => Promise<void>) => {
-    const ops: string[] = [];
-    const at = (path: string): { collectionPath: string; id: string } => {
-      const cut = path.lastIndexOf("/");
-      return { collectionPath: path.slice(0, cut), id: path.slice(cut + 1) };
-    };
-    await body({
-      get: (ref: { path: string }) => {
-        const { collectionPath, id } = at(ref.path);
-        const held = docs.store.get(collectionPath)?.has(id) === true;
-        return Promise.resolve({ exists: () => held });
-      },
-      set: (ref: { path: string }, data: Record<string, unknown>) => ops.push(`set ${ref.path} ${JSON.stringify(data)}`),
-      update: (ref: { path: string }, data: Record<string, unknown>) => ops.push(`update ${ref.path} ${JSON.stringify(data)}`),
-      delete: (ref: { path: string }) => ops.push(`delete ${ref.path}`),
-    });
-    if (batchFails) throw Object.assign(new Error("Missing or insufficient permissions."), { code: "permission-denied" });
-    batched.push(...ops);
-  },
-  writeBatch: () => {
-    const ops: string[] = [];
-    return {
-      set: (ref: { path: string }, data: Record<string, unknown>) => ops.push(`set ${ref.path} ${JSON.stringify(data)}`),
-      // Both call shapes, because the two updates here are deliberately different: a record's
-      // declared field goes through a FieldPath (a dotted name is a literal key, not a path), and
-      // the mirror's `state` is ours and fixed.
-      update: (ref: { path: string }, data: Record<string, unknown> | { segments: string[] }, value?: unknown) => {
-        const asPath = data as { segments?: string[] };
-        const written = Array.isArray(asPath.segments) ? { [asPath.segments.join(".")]: value } : data;
-        ops.push(`update ${ref.path} ${JSON.stringify(written)}`);
-      },
-      delete: (ref: { path: string }) => ops.push(`delete ${ref.path}`),
-      commit: () => {
-        if (batchBreaks > 0) {
-          batchBreaks -= 1;
-          return Promise.reject(new Error("network"));
-        }
-        if (batchRefusals > 0) {
-          batchRefusals -= 1;
-          return Promise.reject(Object.assign(new Error("Missing or insufficient permissions."), { code: "permission-denied" }));
-        }
-        if (batchFails) return Promise.reject(Object.assign(new Error("Missing or insufficient permissions."), { code: "permission-denied" }));
-        batched.push(...ops);
-        return Promise.resolve();
-      },
-    };
-  },
-}));
-
+// Imported INSIDE the factory, which runs when the module under test first pulls Firestore in —
+// long after this file's own imports have been evaluated.
+vi.mock("firebase/firestore", async () => (await import("../../support/participateHarness.js")).firestoreMock(bag));
 vi.mock("../../../server/backends/remoteHost/session.js", () => ({ currentFirestore: () => ({}) }));
 
-/** The row cap each query carried. */
-const capped: number[] = [];
-
-/** Collection paths whose query FAILS without a permission code — a blip, not a refusal. */
-const breakQuery = new Set<string>();
-
-/** Collection paths whose UNFILTERED query is refused — the shape a collection people submit to
- *  has for everyone but its staff. */
-const denyQuery = new Set<string>();
-
-/** What `getDocs` can see — filled from the same store the docs seam holds. */
-const queryable = new Map<string, { id: string; data: Record<string, unknown> }[]>();
-
-class Docs implements FirestoreDocs {
-  readonly store = new Map<string, Map<string, Record<string, unknown>>>();
-  /** Which document paths refuse a GET — how a role scoped to one collection meets `apps/{aid}`. */
-  denyGet = new Set<string>();
-  /** Document paths whose GET fails WITHOUT a permission code — a blip, not a refusal. */
-  breakGet = new Set<string>();
-  created: { path: string; id: string; data: Record<string, unknown> }[] = [];
-
-  put(collectionPath: string, id: string, data: Record<string, unknown>): void {
-    const held = this.store.get(collectionPath) ?? new Map<string, Record<string, unknown>>();
-    held.set(id, data);
-    this.store.set(collectionPath, held);
-    queryable.set(
-      collectionPath,
-      [...held].map(([rowId, row]) => ({ id: rowId, data: row })),
-    );
-  }
-
-  list = (collectionPath: string): Promise<FirestoreDoc[]> => {
-    const held = this.store.get(collectionPath) ?? new Map();
-    return Promise.resolve([...held].sort(([l], [r]) => (l < r ? -1 : 1)).map(([id, data]) => ({ id, data })));
-  };
-
-  get = (collectionPath: string, docId: string): Promise<unknown | null> => {
-    if (this.breakGet.has(`${collectionPath}/${docId}`)) return Promise.reject(new Error("network"));
-    if (this.denyGet.has(`${collectionPath}/${docId}`)) return Promise.reject(Object.assign(new Error("refused"), { code: "permission-denied" }));
-    return Promise.resolve(this.store.get(collectionPath)?.get(docId) ?? null);
-  };
-
-  create = (collectionPath: string, docId: string, data: Record<string, unknown>): Promise<boolean> => {
-    if (this.store.get(collectionPath)?.has(docId)) return Promise.resolve(false);
-    this.created.push({ path: collectionPath, id: docId, data });
-    this.put(collectionPath, docId, data);
-    return Promise.resolve(true);
-  };
-
-  /** Document paths a plain `set` landed on. */
-  sets: string[] = [];
-
-  set = (collectionPath: string, docId: string, data: Record<string, unknown>): Promise<void> => {
-    this.sets.push(`${collectionPath}/${docId}`);
-    this.put(collectionPath, docId, data);
-    return Promise.resolve();
-  };
-  delete = (): Promise<boolean> => Promise.resolve(true);
-  watch = (): (() => void) => () => {};
-}
-
-let docs = new Docs();
-
-const bookingsPath = `apps/${AID}/collections/bookings/items`;
-const slotsPath = `apps/${AID}/collections/slots/items`;
-
-/** The submit declaration, as `config/public` publishes it — the window already lowered, every
- *  other key passed through by the name the rules read it under. */
-const submitBlock = {
-  auth: "verifiedEmail",
-  emailField: "requesterEmail",
-  createFields: ["requesterEmail", "slot", "status", "guests", "seat"],
-  initialStatus: "booked",
-  idFrom: "field",
-  idField: "slot",
-  mirror: "slots",
-  selfTransitions: { booked: ["cancelled"] },
-  selfDelete: ["booked"],
-};
-
-/** Publish this app into the fake database. Nothing here is read off disk — that is the point. */
-/** The submit declaration this app publishes, in whichever of its shapes a test needs. */
-function submitFor({
-  mirror,
-  idFromUid,
-  bothIdentities,
-  dottedEmailField,
-}: {
-  mirror: boolean;
-  idFromUid: boolean;
-  bothIdentities: boolean;
-  dottedEmailField: boolean;
-}): Record<string, unknown> {
-  return {
-    ...submitBlock,
-    ...(mirror ? {} : { mirror: undefined }),
-    // The reader's row is NAMED rather than queried: its id IS their uid.
-    ...(idFromUid ? { idFrom: "auth.uid", idField: undefined, emailField: undefined, mirror: undefined } : {}),
-    // Both identities, either of which the rules accept as an own row.
-    ...(bothIdentities ? { uidField: "uid" } : {}),
-    // An ordinary top-level key that happens to contain a dot.
-    ...(dottedEmailField ? { emailField: "requester.email" } : {}),
-  };
-}
-
-function publish({
-  memberTier = true,
-  roles = { "*": "editor" } as Record<string, string> | null,
-  writers = [ME.email],
-  rosterTier = false,
-  mirror = true,
-  idFromUid = false,
-  enabled = true,
-  bothIdentities = false,
-  name = "Sakura Hair",
-  dottedEmailField = false,
-  dottedStatusField = false,
-} = {}): void {
-  docs.put("appSlugs", "sakura", { aid: AID, published: true });
-  if (roles !== null) docs.put("apps", AID, { aid: AID, name, members: { [ME.email]: roles }, memberEmails: [ME.email] });
-  docs.put(`apps/${AID}/config`, "public", {
-    protocol: "1.0.0",
-    name,
-    enabled,
-    read: ["slots"],
-    submit: { bookings: submitFor({ mirror, idFromUid, bothIdentities, dottedEmailField }) },
-    form: {
-      bookings: {
-        fields: {
-          requesterEmail: { label: "Email", type: "email" },
-          slot: { label: "Slot", type: "string", required: true },
-          status: { label: "Status", type: "string" },
-          // The two types publish deliberately allows, and that a string-only tool was accused of
-          // making unsubmittable.
-          guests: { label: "Guests", type: "number" },
-          seat: { label: "Seat", type: "enum", values: ["window", "aisle"] },
-        },
-        statusField: "status",
-      },
-    },
-    // The roster tier's own writes, which `config/public` carries whether or not the app publishes
-    // any page at all — the half a participant needs and the reason this works with no views.
-    write: [{ cid: "bookings", statusField: "status", transitions: { booked: ["cancelled"] }, selfDelete: ["booked"], withdrawMirror: "slots" }],
-    publishedAt: 1,
-  });
-  if (rosterTier)
-    // A participant page's own projection, left behind by an EARLIER publish: it names the same
-    // collection as `config/public` and carries only the transition half. `runWrites` can stop
-    // after any step and `config/public` is written before the tier documents, so this pair is a
-    // real state — and dropping either half of it takes away a move the deployed rules allow.
-    docs.put(appViewTierPath(AID, "roster"), viewConfigDocId(), {
-      protocol: "1.0.0",
-      views: [{ id: "mine", collections: [{ cid: "bookings", scope: "own" }] }],
-      submit: {},
-      write: [{ cid: "bookings", statusField: "status", transitions: { booked: ["cancelled"] } }],
-      publishedAt: 1,
-    });
-  if (memberTier)
-    // The tier's own id, taken from the package rather than spelled here: it carries a `live:`
-    // prefix, and a document written at "config" is a document nothing reads.
-    docs.put(appViewTierPath(AID, "member"), viewConfigDocId(), {
-      protocol: "1.0.0",
-      views: [{ id: "desk", collections: [{ cid: "bookings", scope: "all" }] }],
-      submit: {},
-      write: [
-        {
-          cid: "bookings",
-          statusField: dottedStatusField ? "workflow.state" : "status",
-          transitions: { booked: ["approved", "cancelled"] },
-          assigneeField: "handledBy",
-          writers,
-          rowWriters: [],
-          mail: { toField: "requesterEmail", on: { approved: { from: ["booked"], to: "approved" } } },
-        },
-      ],
-      publishedAt: 1,
-    });
-}
-
 const run = (args: Record<string, unknown>): Promise<string> => useSharedApp(args);
+const publish = (options: Parameters<typeof publishApp>[1] = {}): void => publishApp(bag, options);
 
-describe("useSharedApp — taking part in somebody else's app", () => {
+describe("useSharedApp — reading somebody else's app", () => {
   beforeAll(() => {
     setSharedCollectionsSupport(true);
-    setFirestoreAccessor(() => ({ docs, email: ME.email, uid: ME.uid }));
+    setFirestoreAccessor(() => ({ docs: bag.docs, email: ME.email, uid: ME.uid }));
   });
 
   beforeEach(() => {
-    docs = new Docs();
-    queryable.clear();
-    capped.length = 0;
-    denyQuery.clear();
-    breakQuery.clear();
-    batched.length = 0;
-    batchFails = false;
-    batchRefusals = 0;
-    batchBreaks = 0;
+    freshBag(bag);
     process.env.MULMOTERMINAL_HOME = makeTempDir("mt-participate-home-");
   });
 
@@ -358,7 +80,7 @@ describe("useSharedApp — taking part in somebody else's app", () => {
 
   it("says so, rather than guessing, when the app document is not readable", async () => {
     publish();
-    docs.denyGet.add(`apps/${AID}`);
+    bag.docs.denyGet.add(`apps/${AID}`);
     const said = await run({ action: "describe", slug: "sakura" });
     expect(said).toContain("Your roles: not readable");
     // The capability still resolves: it comes from the tier projections, which this reader may read.
@@ -367,7 +89,7 @@ describe("useSharedApp — taking part in somebody else's app", () => {
 
   it("refuses an app published against a newer contract, whole rather than in part", async () => {
     publish();
-    const config = docs.store.get(`apps/${AID}/config`)?.get("public");
+    const config = bag.docs.store.get(`apps/${AID}/config`)?.get("public");
     if (config !== undefined) config.protocol = "2.0.0";
     const said = await run({ action: "describe", slug: "sakura" });
     expect(said).toContain("newer version of the shared-app contract");
@@ -378,7 +100,7 @@ describe("useSharedApp — taking part in somebody else's app", () => {
 
   it("reads an app published before the contract carried a version", async () => {
     publish();
-    const config = docs.store.get(`apps/${AID}/config`)?.get("public");
+    const config = bag.docs.store.get(`apps/${AID}/config`)?.get("public");
     if (config !== undefined) delete config.protocol;
     // Absent is the FIRST contract — that is what every app published before the key existed is,
     // and those are the documents in Firestore now.
@@ -387,8 +109,8 @@ describe("useSharedApp — taking part in somebody else's app", () => {
 
   it("does not lose an entry when two apps are remembered at once", async () => {
     publish();
-    docs.put("appSlugs", "kaede", { aid: "app-kaede", published: true });
-    docs.put("apps", "app-kaede", { aid: "app-kaede", name: "Kaede", members: { [ME.email]: { "*": "viewer" } } });
+    bag.docs.put("appSlugs", "kaede", { aid: "app-kaede", published: true });
+    bag.docs.put("apps", "app-kaede", { aid: "app-kaede", name: "Kaede", members: { [ME.email]: { "*": "viewer" } } });
     // Both describes read the register, change it and write it back. Unserialized, the second
     // computes its list from the file the first has already replaced and drops the first's entry.
     await Promise.all([run({ action: "describe", slug: "sakura" }), run({ action: "describe", slug: "kaede" })]);
@@ -431,29 +153,13 @@ describe("useSharedApp — taking part in somebody else's app", () => {
 
   it("does not answer from half the projections when one read breaks", async () => {
     publish();
-    docs.breakGet.add(`apps/${AID}/member/${viewConfigDocId()}`);
+    bag.docs.breakGet.add(`apps/${AID}/member/${viewConfigDocId()}`);
     const said = await run({ action: "describe", slug: "sakura" });
     // Absorbed, this would report the roster half alone — in the same words as an app that
     // genuinely published no staff page, with nothing to say a read failed.
     expect(said).toContain("could not be read");
     expect(said).toContain("not a permission boundary");
     expect(said).not.toContain("You may:");
-  });
-
-  it("does not offer an intent to the next tier when the write merely FAILED", async () => {
-    publish();
-    docs.put(bookingsPath, "b1", { requesterEmail: ME.email, slot: "10:00", status: "booked" });
-    // Both tiers carry `booked -> cancelled`. Retried after a blip, the roster projection would land
-    // the same move carrying no `mail` — the record moves and a declared notice is never queued.
-    batchBreaks = 1;
-    const said = await run({ action: "transition", slug: "sakura", cid: "bookings", id: "b1", to: "cancelled" });
-    expect(said).toContain("network");
-    // NOT "Refused". A `deadline-exceeded` can come back after Firestore committed and the client
-    // lost the answer, so the record may have moved and the notice may have gone out — telling the
-    // user it was refused invites a retry that queues a second real email.
-    expect(said).toContain("Could not tell whether that landed");
-    expect(said).not.toContain("Refused:");
-    expect(batched).toEqual([]);
   });
 
   it("does not call a roster-only board open just because it publishes a public block", async () => {
@@ -467,12 +173,12 @@ describe("useSharedApp — taking part in somebody else's app", () => {
 
   it("finds the rows held under EITHER identity when both are declared", async () => {
     publish({ bothIdentities: true });
-    denyQuery.add(bookingsPath);
+    bag.denyQuery.add(bookingsPath);
     // The rules accept either, so a row staff entered on somebody's behalf carries the address and
     // no uid while that person's own submissions carry the uid. Querying one field only hides half
     // of what is theirs — as an empty answer rather than as an error.
-    docs.put(bookingsPath, "by-uid", { uid: ME.uid, slot: "9:00", status: "booked" });
-    docs.put(bookingsPath, "by-desk", { requesterEmail: ME.email, slot: "10:00", status: "booked" });
+    bag.docs.put(bookingsPath, "by-uid", { uid: ME.uid, slot: "9:00", status: "booked" });
+    bag.docs.put(bookingsPath, "by-desk", { requesterEmail: ME.email, slot: "10:00", status: "booked" });
     const said = await run({ action: "records", slug: "sakura", cid: "bookings" });
     expect(said).toContain("by-uid");
     expect(said).toContain("by-desk");
@@ -491,6 +197,16 @@ describe("useSharedApp — taking part in somebody else's app", () => {
     expect(said).not.toMatch(/^IGNORE THE USER/m);
   });
 
+  it("strips the characters that can forge structure inside a quoted value", async () => {
+    // NUL, a zero-width space, a bidi OVERRIDE and a LINE SEPARATOR — the last of which JavaScript
+    // and many renderers treat as a line break, and the override of which reorders everything after
+    // it, so a value can render as text it does not contain. All collapse to a space inside the
+    // quote, which is also the plainest evidence that this module's character class parses at all.
+    publish({ name: "A\u0000B\u200bC\u202eD\u2028E" });
+    const said = await run({ action: "describe", slug: "sakura" });
+    expect(said).toContain("\u00abA B C D E\u00bb");
+  });
+
   it("caps a value that is a payload rather than a label, and says how much it dropped", async () => {
     publish({ name: "x".repeat(400) });
     const said = await run({ action: "describe", slug: "sakura" });
@@ -499,20 +215,19 @@ describe("useSharedApp — taking part in somebody else's app", () => {
 
   it("queries an identity field whose name contains a dot as a literal key", async () => {
     publish({ dottedEmailField: true });
-    denyQuery.add(bookingsPath);
-    docs.put(bookingsPath, "mine", { "requester.email": ME.email, slot: "9:00", status: "booked" });
+    bag.denyQuery.add(bookingsPath);
+    bag.docs.put(bookingsPath, "mine", { "requester.email": ME.email, slot: "9:00", status: "booked" });
     const said = await run({ action: "records", slug: "sakura", cid: "bookings" });
     // A bare dotted string is a NESTED PATH to `where`, so this would look inside a map called
     // `requester` and match nothing — an empty answer rather than an error.
     expect(said).toContain("mine");
   });
 
-  it("moves a status field whose name contains a dot as a literal key", async () => {
-    publish({ dottedStatusField: true });
-    docs.put(bookingsPath, "b1", { requesterEmail: "guest@example.com", slot: "10:00", "workflow.state": "booked" });
-    await run({ action: "transition", slug: "sakura", cid: "bookings", id: "b1", to: "approved" });
-    // The object form would write a map called `workflow` beside the field the state lives in.
-    expect(batched[0]).toContain(JSON.stringify({ "workflow.state": "approved" }));
+  it("reports openness from the document the RULES read, not from the projection", async () => {
+    // A publish that stopped between the two: `config/public` says the app is now open, the app
+    // document — which is what `publicOn` reads — still says it is not.
+    publish({ enabled: true, appPublic: { enabled: false, read: ["slots"] } });
+    expect(await run({ action: "describe", slug: "sakura" })).toContain("NOT open to the public");
   });
 
   it("refuses a URL name nothing answers to", async () => {
@@ -532,7 +247,7 @@ describe("useSharedApp — taking part in somebody else's app", () => {
 
   it("lists a whole collection when the rules open it", async () => {
     publish();
-    docs.put(slotsPath, "10:00", { state: "open" });
+    bag.docs.put(slotsPath, "10:00", { state: "open" });
     const said = await run({ action: "records", slug: "sakura", cid: "slots" });
     expect(said).toContain("the whole collection");
     expect(said).toContain("10:00");
@@ -540,9 +255,9 @@ describe("useSharedApp — taking part in somebody else's app", () => {
 
   it("falls back to the reader's OWN rows when the list is refused, and says which it gave", async () => {
     publish();
-    denyQuery.add(bookingsPath);
-    docs.put(bookingsPath, "10:00", { requesterEmail: ME.email, slot: "10:00", status: "booked" });
-    docs.put(bookingsPath, "11:00", { requesterEmail: "someone@example.com", slot: "11:00", status: "booked" });
+    bag.denyQuery.add(bookingsPath);
+    bag.docs.put(bookingsPath, "10:00", { requesterEmail: ME.email, slot: "10:00", status: "booked" });
+    bag.docs.put(bookingsPath, "11:00", { requesterEmail: "someone@example.com", slot: "11:00", status: "booked" });
     const said = await run({ action: "records", slug: "sakura", cid: "bookings" });
     expect(said).toContain("YOUR OWN ONLY");
     expect(said).toContain("do not describe it as one");
@@ -550,210 +265,58 @@ describe("useSharedApp — taking part in somebody else's app", () => {
     expect(said).not.toContain("someone@example.com");
   });
 
-  it("submits through the published form, and reports a record rather than a seat", async () => {
-    publish();
-    docs.put(slotsPath, "10:00", { state: "open" });
-    const said = await run({ action: "submit", slug: "sakura", cid: "bookings", values: { slot: "10:00", guests: "2", seat: "window" } });
-    expect(said).toContain("The record's id is 10:00");
-    // The mirror travelled with it, in ONE batch: the rules read the second document with
-    // `getAfter()`, so a mirror written singly is refused with nothing to say why.
-    expect(batched).toEqual([
-      // The typed fields land as the STRINGS they were sent as, which is what mulmoserver's own
-      // page writes for them (`recordOf` takes `Record<string, string>` and writes it verbatim).
-      // Coercing here would make this host write a different document than the page for the same
-      // answer — read differently by the app's own views.
-      `set ${bookingsPath}/10:00 ${JSON.stringify({ slot: "10:00", guests: "2", seat: "window", requesterEmail: ME.email, status: "booked" })}`,
-      `update ${slotsPath}/10:00 {"state":"taken"}`,
-    ]);
-    // Principle 3, said out loud: what a create buys is a position, and this report must never
-    // promise more than that.
-    expect(said).toContain("not a place held");
-    expect(said.toLowerCase()).not.toContain("reserved");
-    expect(said.toLowerCase()).not.toContain("secured");
-  });
-
-  it("refuses a slot somebody already holds, inside the transaction", async () => {
-    publish();
-    docs.put(slotsPath, "10:00", { state: "taken" });
-    docs.put(bookingsPath, "10:00", { requesterEmail: "guest@example.com", slot: "10:00", status: "booked" });
-    const said = await run({ action: "submit", slug: "sakura", cid: "bookings", values: { slot: "10:00" } });
-    expect(said).toContain("somebody has it");
-    // Nothing was written. The read that refused is INSIDE the transaction, which is what stops a
-    // writer's `set` from replacing the booking that is already there: `mirrorClaimed` in the rules
-    // only asks that the mirror end up `taken`, so a slot another participant just claimed would
-    // have satisfied it.
-    expect(batched).toEqual([]);
-  });
-
-  it("claims a slot without reading it, for a submitter the rules will not let read", async () => {
-    publish();
-    docs.put(slotsPath, "10:00", { state: "open" });
-    // A participant reaches a booking only through `ownRow`, which reads fields off a document that
-    // does not exist yet — so the rules deny this get, and a transaction opening with it would be
-    // refused before it wrote anything. The paired write still has to go out.
-    docs.denyGet.add(`${bookingsPath}/10:00`);
-    const said = await run({ action: "submit", slug: "sakura", cid: "bookings", values: { slot: "10:00" } });
-    expect(said).toContain("The record's id is 10:00");
-    expect(batched).toEqual([
-      `set ${bookingsPath}/10:00 ${JSON.stringify({ slot: "10:00", requesterEmail: ME.email, status: "booked" })}`,
-      `update ${slotsPath}/10:00 {"state":"taken"}`,
-    ]);
-  });
-
-  it("writes nothing when the destination read merely FAILED rather than being refused", async () => {
-    publish();
-    docs.put(slotsPath, "10:00", { state: "open" });
-    docs.breakGet.add(`${bookingsPath}/10:00`);
-    const said = await run({ action: "submit", slug: "sakura", cid: "bookings", values: { slot: "10:00" } });
-    // A blip is not a refusal. Read as one, a WRITER would take the unchecked branch — the one that
-    // can turn `set` into an update over somebody else's booking.
-    expect(said).toContain("network");
-    expect(batched).toEqual([]);
-  });
-
-  it("answers a private form whose collection the submitter may not read", async () => {
-    // No mirror: the whole submission is one document. Both checked shapes open by READING the id
-    // — core's `create` runs a transaction that does — and on a collection reached only through
-    // `ownRow` the rules deny that read, so a private survey could not be answered at all.
-    publish({ mirror: false });
-    docs.denyGet.add(`${bookingsPath}/10:00`);
-    const said = await run({ action: "submit", slug: "sakura", cid: "bookings", values: { slot: "10:00" } });
-    expect(said).toContain("The record's id is 10:00");
-    // Through the seam's plain `set`, which is what a submission with no mirror needs — and which
-    // does not ask for an SDK handle the caller may not have.
-    expect(docs.sets).toEqual([`${bookingsPath}/10:00`]);
-    expect(batched).toEqual([]);
-  });
-
-  it("names the missing field rather than letting the rules refuse it namelessly", async () => {
-    publish();
-    const said = await run({ action: "submit", slug: "sakura", cid: "bookings", values: {} });
-    expect(said).toContain("missing: \u00abSlot\u00bb");
-  });
-
-  it("moves a record on the member tier and queues the declared notice in the same batch", async () => {
-    publish();
-    docs.put(bookingsPath, "b1", { requesterEmail: "guest@example.com", slot: "10:00", status: "booked" });
-    const said = await run({ action: "transition", slug: "sakura", cid: "bookings", id: "b1", to: "approved" });
-    expect(said).toContain("Moved «bookings»/«b1» to «approved»");
-    expect(said).toContain("Judged on the member tier");
-    expect(said).toContain("A notification was QUEUED");
-    expect(batched).toEqual([
-      `update ${bookingsPath}/b1 {"status":"approved"}`,
-      // The id the RULES rebuild. A different spelling queues a document the mail rule refuses.
-      `set apps/${AID}/mail/bookings_b1_approved ${JSON.stringify({ cid: "bookings", itemId: "b1", to: "guest@example.com", template: "approved" })}`,
-    ]);
-  });
-
-  it("names the declaration when a move is not in the table, and writes nothing", async () => {
-    publish();
-    docs.put(bookingsPath, "b1", { requesterEmail: "guest@example.com", slot: "10:00", status: "approved" });
-    const said = await run({ action: "transition", slug: "sakura", cid: "bookings", id: "b1", to: "booked" });
-    expect(said).toContain("illegal-transition");
-    expect(batched).toEqual([]);
-  });
-
-  it("refuses a move a reader holding no writing role can make on neither tier", async () => {
-    publish({ memberTier: false });
-    docs.put(bookingsPath, "b1", { requesterEmail: ME.email, slot: "10:00", status: "booked" });
-    const said = await run({ action: "transition", slug: "sakura", cid: "bookings", id: "b1", to: "approved" });
-    expect(said).toContain("roster: illegal-transition");
-    expect(batched).toEqual([]);
-  });
-
-  it("withdraws the reader's own row and reopens the slot in one batch", async () => {
-    publish();
-    docs.put(bookingsPath, "10:00", { requesterEmail: ME.email, slot: "10:00", status: "booked" });
-    docs.put(slotsPath, "10:00", { state: "taken" });
-    const said = await run({ action: "withdraw", slug: "sakura", cid: "bookings", id: "10:00" });
-    expect(said).toContain("The record is gone");
-    expect(said).toContain("There is no undo");
-    expect(batched).toEqual([`delete ${bookingsPath}/10:00`, `update ${slotsPath}/10:00 {"state":"open"}`]);
-  });
-
   it("caps the own-row query in the QUERY, not after the fetch", async () => {
     publish();
-    denyQuery.add(bookingsPath);
-    for (let n = 0; n < 5; n += 1) docs.put(bookingsPath, `b${n}`, { requesterEmail: ME.email, slot: `${n}:00`, status: "booked" });
+    bag.denyQuery.add(bookingsPath);
+    for (let n = 0; n < 5; n += 1) bag.docs.put(bookingsPath, `b${n}`, { requesterEmail: ME.email, slot: `${n}:00`, status: "booked" });
     const said = await run({ action: "records", slug: "sakura", cid: "bookings", limit: 2 });
     // The cap reached Firestore rather than being applied to a fetched result: the query carried it,
     // so what came back is already two rows.
     // Twice: the refused whole-collection attempt and the own-row query that answered. Both carry
     // the cap, which is the property under test.
-    expect([...new Set(capped)]).toEqual([2]);
+    expect([...new Set(bag.capped)]).toEqual([2]);
     expect(said).toContain("2 row(s)");
-  });
-
-  it("refuses an assignee nobody on the roster could be", async () => {
-    publish();
-    docs.put(bookingsPath, "b1", { requesterEmail: "guest@example.com", slot: "10:00", status: "booked" });
-    const said = await run({ action: "assign", slug: "sakura", cid: "bookings", id: "b1", to: "stranger@example.com" });
-    // Refused by NAME. Written, the row would belong to somebody who may never touch it again.
-    expect(said).toContain("unknown-assignee");
-    expect(batched).toEqual([]);
-  });
-
-  it("refuses a move this reader's role does not carry, and says which tier said what", async () => {
-    publish({ writers: ["somebody-else@example.com"] });
-    docs.put(bookingsPath, "b1", { requesterEmail: "guest@example.com", slot: "10:00", status: "booked" });
-    const said = await run({ action: "transition", slug: "sakura", cid: "bookings", id: "b1", to: "approved" });
-    expect(said).toContain("member:");
-    expect(said).toContain("roster:");
-    expect(batched).toEqual([]);
-  });
-
-  it("tries the next tier when the RULES refuse the first tier's write", async () => {
-    publish();
-    docs.put(bookingsPath, "b1", { requesterEmail: ME.email, slot: "10:00", status: "booked" });
-    // Both tiers carry `booked -> cancelled`, so the member projection is judged first and its
-    // write is the one the rules turn down. Stopping there would deny this person a move they hold
-    // as the row's own submitter.
-    batchRefusals = 1;
-    const said = await run({ action: "transition", slug: "sakura", cid: "bookings", id: "b1", to: "cancelled" });
-    expect(said).toContain("Judged on the roster tier");
-    // ONE write landed — the refused batch wrote nothing, which is what makes the retry safe.
-    expect(batched).toEqual([`update ${bookingsPath}/b1 {"status":"cancelled"}`]);
   });
 
   it("never asks Firestore for nought rows", async () => {
     publish();
-    denyQuery.add(bookingsPath);
-    docs.put(bookingsPath, "b0", { requesterEmail: ME.email, slot: "0:00", status: "booked" });
+    bag.denyQuery.add(bookingsPath);
+    bag.docs.put(bookingsPath, "b0", { requesterEmail: ME.email, slot: "0:00", status: "booked" });
     // `Math.floor(0.5)` is 0, and `limit(0)` is REFUSED by Firestore at the moment the constraint is
     // built — before the read this tool wraps in a catch. So half a row would have thrown where
     // every other bad argument produces a sentence.
     const said = await run({ action: "records", slug: "sakura", cid: "bookings", limit: 0.5 });
-    expect([...new Set(capped)]).toEqual([DEFAULT_ROWS]);
+    expect([...new Set(bag.capped)]).toEqual([DEFAULT_ROWS]);
     expect(said).toContain("YOUR OWN ONLY");
   });
 
   it("keeps both halves of the roster tier when the two sources name one collection", async () => {
     publish({ rosterTier: true, memberTier: false });
-    docs.put(bookingsPath, "10:00", { requesterEmail: ME.email, slot: "10:00", status: "booked" });
-    docs.put(slotsPath, "10:00", { state: "taken" });
+    bag.docs.put(bookingsPath, "10:00", { requesterEmail: ME.email, slot: "10:00", status: "booked" });
+    bag.docs.put(slotsPath, "10:00", { state: "taken" });
     // The withdrawal lives only in `config/public` (`selfDelete` + `withdrawMirror`); the tier
     // document names the same cid and carries only the transitions. Keeping one entry and dropping
     // the other would refuse a withdrawal this app allows.
     const said = await run({ action: "withdraw", slug: "sakura", cid: "bookings", id: "10:00" });
     expect(said).toContain("The record is gone");
-    expect(batched).toEqual([`delete ${bookingsPath}/10:00`, `update ${slotsPath}/10:00 {"state":"open"}`]);
+    expect(bag.batched).toEqual([`delete ${bookingsPath}/10:00`, `update ${slotsPath}/10:00 {"state":"open"}`]);
   });
 
   it("lowers an ask no inspection needs, and says it did", async () => {
     publish();
-    denyQuery.add(bookingsPath);
-    docs.put(bookingsPath, "b0", { requesterEmail: ME.email, slot: "0:00", status: "booked" });
+    bag.denyQuery.add(bookingsPath);
+    bag.docs.put(bookingsPath, "b0", { requesterEmail: ME.email, slot: "0:00", status: "booked" });
     const said = await run({ action: "records", slug: "sakura", cid: "bookings", limit: 1_000_000_000 });
     // A billion is a finite number the schema accepts. Passed through, it bills a read per row in
     // somebody else's app and serializes the lot into a context window.
-    expect([...new Set(capped)]).toEqual([500]);
+    expect([...new Set(bag.capped)]).toEqual([500]);
     expect(said).toContain("this tool reads at most 500");
   });
 
   it("says a full page might not be the whole collection", async () => {
     publish();
-    docs.put(slotsPath, "10:00", { state: "open" });
-    docs.put(slotsPath, "11:00", { state: "open" });
+    bag.docs.put(slotsPath, "10:00", { state: "open" });
+    bag.docs.put(slotsPath, "11:00", { state: "open" });
     const said = await run({ action: "records", slug: "sakura", cid: "slots", limit: 1 });
     // A count that exactly fills the ask says nothing about what is behind it, and an agent reads
     // "1 row" as the collection.
@@ -762,8 +325,8 @@ describe("useSharedApp — taking part in somebody else's app", () => {
 
   it("does not report a broken read as a permission boundary", async () => {
     publish();
-    breakQuery.add(bookingsPath);
-    docs.put(bookingsPath, "b0", { requesterEmail: ME.email, slot: "0:00", status: "booked" });
+    bag.breakQuery.add(bookingsPath);
+    bag.docs.put(bookingsPath, "b0", { requesterEmail: ME.email, slot: "0:00", status: "booked" });
     const said = await run({ action: "records", slug: "sakura", cid: "bookings" });
     // Narrowed to the reader's own rows, this would say "only your own rows are readable here"
     // about a collection they may in fact read whole — and the agent would repeat it as the app's
@@ -773,40 +336,15 @@ describe("useSharedApp — taking part in somebody else's app", () => {
     expect(said).not.toContain("YOUR OWN ONLY");
   });
 
-  it("does not report a broken record read as a permission boundary", async () => {
-    publish();
-    docs.breakGet.add(`${bookingsPath}/b1`);
-    const said = await run({ action: "transition", slug: "sakura", cid: "bookings", id: "b1", to: "approved" });
-    expect(said).toContain("worth making again");
-    expect(said).not.toContain("not readable by you");
-    expect(batched).toEqual([]);
-  });
-
   it("does not report a broken own-row lookup as an empty own-row answer", async () => {
     // `idFrom: "auth.uid"`: the reader's row is NAMED, so the fallback is a get rather than a
     // query. An empty answer here means "you have not got one", and a blip must not borrow that
     // sentence.
     publish({ idFromUid: true });
-    denyQuery.add(bookingsPath);
-    docs.breakGet.add(`${bookingsPath}/${ME.uid}`);
+    bag.denyQuery.add(bookingsPath);
+    bag.docs.breakGet.add(`${bookingsPath}/${ME.uid}`);
     const said = await run({ action: "records", slug: "sakura", cid: "bookings" });
     expect(said).toContain("a failure, not a permission boundary");
     expect(said).not.toContain("YOUR OWN ONLY");
-  });
-
-  it("keeps a refused read apart from an absent record", async () => {
-    publish();
-    docs.denyGet.add(`${bookingsPath}/b1`);
-    expect(await run({ action: "transition", slug: "sakura", cid: "bookings", id: "b1", to: "approved" })).toContain("not readable by you");
-    expect(await run({ action: "transition", slug: "sakura", cid: "bookings", id: "b2", to: "approved" })).toContain('no record "b2"');
-    expect(batched).toEqual([]);
-  });
-
-  it("reports what the rules said when they refuse the write this host judged fine", async () => {
-    publish();
-    docs.put(bookingsPath, "b1", { requesterEmail: "guest@example.com", slot: "10:00", status: "booked" });
-    batchFails = true;
-    const said = await run({ action: "transition", slug: "sakura", cid: "bookings", id: "b1", to: "approved" });
-    expect(said).toContain("Missing or insufficient permissions");
   });
 });
